Move sheet help view into a more general-purpose help view that can be used by other future help windows.
